+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 
 - Add test for `maxLintingIssues` setting
+- Add test for `lintingEnabled` setting
 
 ## [2.0.3] - 2025-12-16